+++ src/dipdup/report.py
@@ -9,6 +9,7 @@ from dipdup.yaml import dump
 
 # FIXME: Hardcoded path
 REPORTS_PATH = Path.home() / '.local' / 'share' / 'dipdup' / 'reports'
+REPORTS_LIMIT = 100
 
 
 class ReportHeader(TypedDict):
@@ -58,3 +59,10 @@ def save_report(package: str, error: Exception | None) -> str:
     event_yaml = dump(event)
     path.write_text(event_yaml)
     return report_id
+
+
+def cleanup_reports() -> None:
+    """Removes old reports"""
+    for i, path in enumerate(REPORTS_PATH.glob('*.yaml')):
+        if i > 100:
+            path.unlink()